beam/requirements.txt
applicationinsights
backports.csv
certifi
chardet
cli-helpers
Click
configobj
defusedxml
Django
django-allauth
django-cors-headers
django-cors-middleware
django-dotenv
django-extensions
django-mssql
django-redis
django-rest-auth
djangorestframework
docopt
enum34
future
gunicorn
humanize
idna
mandrill
mssql-cli
nltk
oauthlib
prompt-toolkit
Pygments
pymongo
PyMySQL
python-openid
python3-openid
redis
requests
requests-oauthlib
six
sqlparse
terminaltables
Unipath
urllib3
wcwidth
xmltodict
